Refactor twitter action
